refactor(utils,apis_entities): move `utils.rdf` to `apis_entities.utils`
The `rdf` module is only used to import entities, so it makes sense to
put it in the module where it is used.

Closes: #2215
