+++ apis_core/apis_entities/utils/rdf.py
